Tab limit for tab groups (#46)
When tab groups reach a certain limit, a new group of that domain is created. This can be disabled in `settings.limitTabGroupSize`. The max number of tabs is defined in `settings.maxTabsPerGroup`

The `TabGroup.id` attribute has been changed, as it cannot be just a transformation of the `hostname`. For now, it gets and id equal to the number of instances of the class created so far.