Release 0.2.0 (#128)
* Updated link to release 0.1.0

Updated link to release 0.1.0

* Refactored project (#45)

* Tab limit for tab groups (#46)

When tab groups reach a certain limit, a new group of that domain is created. This can be disabled in `settings.limitTabGroupSize`. The max number of tabs is defined in `settings.maxTabsPerGroup`

The `TabGroup.id` attribute has been changed, as it cannot be just a transformation of the `hostname`. For now, it gets and id equal to the number of instances of the class created so far.

* Updated the readme (#48)

* Added toggle settings for limitting tabgroup size (#49)

closes #47

* Added settings (#51)

- Closing the tab manager after clicking  a tab is now a setting
- Added a toggle to limit the number of tabs in a tab group
- Added a slider to change the max number of tabs from 1 to 10
closes #18

* Move search options close to search bar (#68)

Moved search options to under the search bar.

* Max tabs per page slider (#71)

The callback to update the max number of tabs from the slider triggers as the user slides without having to release the click button.

* Reopen tabGroups with the Undo button (#73)

* Undo closing tab groups

Tab groups can be reopened after being closed.

Also included `underscore.js` to handle collections and arrays

* Styled radio buttons (#75)

* Added number to max tab group setting slider (#78)

* View all windows on one page (#82)

* Added functionality to group tabs by window (#79)

The view can be changed between the old version and the new version by changing the tab manager setting `'showAllWindowsTogether'`

* Fixed layout and styled all windows (#81)

Styled windows

* Added window number to view all windows section

* Added scroll for overflown all windows section

* Added close window buttons

* Undo now works for closing windows

* Select window buttons work as hrefs (#87)

* Added smooth scrolling for windows (#90)

* Refactor settings (#91)

The ids of the settings that are in a checkbox or in a slider must be included in the `settingsIds` list of the TabManager.

The callbacks for these components will be update automatically from these IDs and also their values will be loaded and stored from the chrome storage.

Added an HTML toggle in the sidebar to change between the classic view mode and the new one

* Added firefox support (#93)

* Firefox - Added HTML encoding and Platform.js (#70)

* HTML encoding and click on tab

- A `<meta>` tag has been added to the `tabPage.html` file to define the encoding
- The `highlighted` property has been changed by `active`

* Platform.js and conditional updateProperties

Added Platform.js, a JavaScript library to get information about the current platform and browser.

Used this library to make a different payload for the call `chrome.tabs.update` depending on the platform name

* Fixed tabs not loading on launch (#77)

* Fix multiple managers open at once (#92)

Fixed bug with multiple managers open at the same time

* Create ISSUE_TEMPLATE.md (#94)

* Arrange tabs now works for all windows (#95)

* Added classic functionality to the 'All' button (#97)

* Styled All section and classic mode (#99)

* Added shortcuts for settings and arrangeTabs (#100)

* Fixed undo opening duplicate tabs (#102)

* Fixed setting bugs (#104)

* Fixed the tabs being filtered by window in the new mode (#107)

* Added active window btn changes on scroll (#109)

* Fixed undo opening tabGroup in multiple windows (#110)

* Fixed tab groups not merging in view all mode (#111)

* Fixed no search results found text is hidden by the window container (#113)

* Some renders in promises and unique IDs for tabgroups and tab containers (#116)

- Added promises to the rendering of the tab groups and the tabs; now, they will be rendered in parallel
- Fixed https://github.com/adamjamesadair/manage-my-tabs/issues/115 by having a different id for tabgroups and for tabs

* Fixed tab group listener ids

* Added color animation when a window is selected (#117)

Created a new CSS and a jQuery animation that will be triggered every time a window is selected by clicking the buttons in "Select Window:"

* Overscroll main container as a setting (#120)

- Added `'allowOverscrollMainContainer'` as a generic setting and solved this issue: https://github.com/adamjamesadair/manage-my-tabs/issues/119
- Removed `console.log` from rendering promises
- Renamed `tabManager.storeSettings` to `tabManager.addGenericSettingsCallbacks`

* Added border and reduced delay in window animation (#121)

* Fixed multiple tabs open after undo of last tab group in window (#123)

* Updated Readme

* Update changelog

* Updated changelogs

* Updated store content

* Updated Readme

* Shortcuts (#125)

Added shorcuts to:
- Select the next and the previous window, taking the window with the `window-select-active` class as reference
- Change to the 'all' mode (Ctrl + 0)

* Update to version 0.2.0

* Updated the Readme